Prometheus: fix scheduler & controller manager monitoring
By default all token requests are considered to be anonymous. As a
result if RBAC is enabled, Prometheus can't scrape the Kubernetes
scheduler nor the controller manager. This change allows both components to
verify the token and authorize the request.
Additionally both components are being exposed to Prometheus by a
service.