[cis-hardening] Fix `--install-kubebench` flag
The `--install-kubebench` flag is defined without a specific type which
let click interpret it as `string`. The check of this flag expected a
boolean (`if install_kubebench:...`) which caused
`--install-kubebench="false"` to be evaluated to `True`.

This commit fixes that issue and evaluates the flag correctly. However,
we can not simply change the flag type to boolean as this would break
the current workaround for this issue (`--install_kubebench=""`). Thus,
we maintain the `string` type for this flag for backward-compatibility.

Relates to MK-1388