minio: Adds additional minio tenant user
Previously, when the minio krew plugin was used to deploy the minio tenant,
the krew plugin created an additional user, with the credentials set in the
$TENANT_NAME-user-1 Kubernetes secret.

The Helm chart we're using does not do that by default, but it does accept
a list of Kubernetes secrets with CONSOLE_ACCESS_KEY (user) and
CONSOLE_SECRET_KEY (password) fields set in them.

With this change, we're now creating the secret ourselves, and passing it
onto the Helm chart. The minio tenant will have this new user, in addition
to the regular root user. With this, the behaviour that the krew plugin had
is preserved.

Additionally, remvoes reference to the kubectl-minio plugin since it doesn't
exist anymore.