Updates minio root user and password
Previously, when the minio krew plugin was used to deploy the minio
tenant, the root user credentials were randomly generated.

The Helm chart we're using have default values: minio, minio123.

We're now generating random strings for the root user and password.
Additionally, we're updating the minio env configuration secret name
to reflect the tenant name, as it was previously.